Fix bool and null values not being auto-skipped in lists and dicts
When iterating over a list or a dict as shown in the following example,
unused values should automatically be skipped.

Code:
    if Reader.Dict then
      while (Reader.Advance <> jsDictEnd) and Reader.Key(k) do
        if k = 'a' then
        begin
          Reader.Str(s)
          WriteLn('a: ', s);
        end
        else if k = 'b' then
        begin
          Reader.Number(n)
          WriteLn('b: ', n);
        end;

Input:
    {
      "a": "Hello",
      "unknown_key": "xyz",
      "b": 42
    }

Output:
    a: Hello
    b: 42

Notice how "unknown_key" is automatically skipped since we did not
explicitly handle it. This is intentional.

However, this behavior was broken if the value of the extra item was a
bool or a null.  In that case, all items after the unknown item would be
skipped. For example, in the following JSON, all items after
"unknown_key" were ignored:

Input:
    {
      "a": "Hello",
      "unknown_key": null,
      "b": 42
    }

Output (old):
    a: Hello

Output (new):
    a: Hello
    b: 42

The same bug also occured inside lists.